requirements/requirements.txt
requests
jenkinsapi
supervisor
click
rpi.gpio
